requirements.txt
Django==2.2.6
django-crispy-forms==1.7.2
django-filter==2.0.0
pytz==2018.5
